requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Migrate==4.1.*
Flask-APScheduler==1.13.*
SQLAlchemy==2.0.*
alembic==1.14.*
psycopg2-binary==2.9.*
gunicorn==23.*
feedparser==6.0.*
readability-lxml==0.8.*
beautifulsoup4==4.12.*
lxml==5.*
requests==2.32.*
openai==1.60.*
numpy==2.*
scikit-learn==1.6.*
yfinance==0.2.*
python-dotenv==1.0.*
pytest==8.*
pytest-flask==1.3.*
